# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 62c88dacc97c70218d659175
  title: captureTest
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
  /login:
    post:
      x-skaffolder-ignore: true
      summary: Login API
      security: []
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              example:
                username: admin
                password: 62f264d7ad826f02a8af714c0a54b197935b717656b80461686d450f7b3abde4c553541515de2052b9af70f710f0cd8a1a2d3f4d60aa72608d71a63a9a93c0f5
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /verifyToken:
    post:
      x-skaffolder-ignore: true
      summary: Verify token API
      security: []
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /changePassword:
    post:
      x-skaffolder-ignore: true
      summary: Change password API
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordOld:
                  type: string
      responses:
        "200":
          description: OK
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      x-skaffolder-id: 62c88dacc97c70218d65917f
      x-skaffolder-id-db: 62c88dacc97c70218d659176
      x-skaffolder-id-entity: 62c88dacc97c70218d659177
      x-skaffolder-url: /user
      x-skaffolder-type: User
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        mail:
          type: string
          x-skaffolder-id-attr: 62c88dacc97c70218d65917a
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 62c88dacc97c70218d65917b
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 62c88dacc97c70218d659179
          x-skaffolder-type: String
          x-skaffolder-required: true
        roles:
          type: string
          x-skaffolder-id-attr: 62c88dacc97c70218d65917d
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 62c88dacc97c70218d65917c
          x-skaffolder-type: String
        username:
          type: string
          x-skaffolder-id-attr: 62c88dacc97c70218d659178
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - username
...